LayoutTests/imported/w3c/web-platform-tests/svg/animations/event-listeners-expected.txt


FAIL When the animation ends, only the 'onend' attribute + IDL listeners and the 'endEvent' listener should be called assert_equals: expected 1 but got 0
